unital_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.7
Pillow==10.0.0
python-decouple==3.8
djangorestframework-simplejwt==5.2.2